refactor nakadi types
