+++ nakadi-types/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nakadi-types"
-version = "0.1.0-beta.1"
+version = "0.1.0"
 authors = ["Christian Douven <chridou@users.noreply.github.com>"]
 readme = "README.md"
 license = "Apache-2.0/MIT"
