+++ nakadi-types/src/event/publishable.rs
@@ -1,8 +1,5 @@
 //! Publishable events
-pub use crate::{
-    model::{event_type::EventTypeName, partition::PartitionId},
-    FlowId,
-};
+pub use crate::{event_type::EventTypeName, partition::PartitionId, FlowId};
 
 use chrono::{DateTime, Utc};
 use serde::Serialize;
