+++ nakadi-types/src/misc.rs
